Feat : 가게 관련 요청 유저 인증 추가 및 예외처리 추가

+++ src/main/java/com/sparta/palpaleats/domain/store/service/StoreService.java
@@ -19,6 +19,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.UnsupportedEncodingException;
 import java.util.List;
+import java.util.Map;
 
 import static com.sparta.palpaleats.domain.user.entity.UserRoleEnum.SELLER;
 
@@ -59,7 +60,13 @@ public class StoreService {
 
 
     @Transactional
-    public CommonResponseDto updatePicture(MultipartFile multipartFile, Long storeId) throws UnsupportedEncodingException {
+    public CommonResponseDto updatePicture(MultipartFile multipartFile, Long storeId, Long id) throws UnsupportedEncodingException {
+        User user = userRepository.findById(id).orElseThrow(
+                () -> new CustomException(ExceptionCode.NOT_FOUND_USER)
+        );
+        if(!user.getRole().equals(SELLER)){
+            throw new CustomException(ExceptionCode.FORBIDDEN_YOUR_NOT_SELLER);
+        }
         Store store = findStore(storeId);
         if (s3Util.validateFileExists(multipartFile)) {
             s3Service.deleteImage(store.getStorePicturePath());
@@ -71,21 +78,39 @@ public class StoreService {
 
 
     @Transactional
-    public CommonResponseDto updateStoreName(String name, Long storeId) {
+    public CommonResponseDto updateStoreName(String name, Long storeId, Long id) {
+        User user = userRepository.findById(id).orElseThrow(
+                () -> new CustomException(ExceptionCode.NOT_FOUND_USER)
+        );
+        if(!user.getRole().equals(SELLER)){
+            throw new CustomException(ExceptionCode.FORBIDDEN_YOUR_NOT_SELLER);
+        }
         Store store = findStore(storeId);
         store.setName(name);
         return new CommonResponseDto(CommonResponseCode.STORE_UPDATE);
     }
 
     @Transactional
-    public CommonResponseDto updateStoreCategory(String category, Long storeId) {
+    public CommonResponseDto updateStoreCategory(String category, Long storeId, Long id) {
+        User user = userRepository.findById(id).orElseThrow(
+                () -> new CustomException(ExceptionCode.NOT_FOUND_USER)
+        );
+        if(!user.getRole().equals(SELLER)){
+            throw new CustomException(ExceptionCode.FORBIDDEN_YOUR_NOT_SELLER);
+        }
         Store store = findStore(storeId);
         store.setCategory(category);
         return new CommonResponseDto(CommonResponseCode.STORE_UPDATE);
     }
 
     @Transactional
-    public CommonResponseDto updateStoreAddress(String address, Long storeId){
+    public CommonResponseDto updateStoreAddress(String address, Long storeId, Long id){
+        User user = userRepository.findById(id).orElseThrow(
+                () -> new CustomException(ExceptionCode.NOT_FOUND_USER)
+        );
+        if(!user.getRole().equals(SELLER)){
+            throw new CustomException(ExceptionCode.FORBIDDEN_YOUR_NOT_SELLER);
+        }
         Store store = findStore(storeId);
         store.setAddress(address);
         return new CommonResponseDto(CommonResponseCode.STORE_UPDATE);
@@ -93,7 +118,13 @@ public class StoreService {
 
 
     @Transactional
-    public CommonResponseDto updateStorePhone(String phone, Long storeId) {
+    public CommonResponseDto updateStorePhone(String phone, Long storeId, Long id) {
+        User user = userRepository.findById(id).orElseThrow(
+                () -> new CustomException(ExceptionCode.NOT_FOUND_USER)
+        );
+        if(!user.getRole().equals(SELLER)){
+            throw new CustomException(ExceptionCode.FORBIDDEN_YOUR_NOT_SELLER);
+        }
         Store store = findStore(storeId);
         store.setPhone(phone);
         return new CommonResponseDto(CommonResponseCode.STORE_UPDATE);
@@ -101,33 +132,69 @@ public class StoreService {
 
 
     @Transactional
-    public CommonResponseDto updateStoreContent(String content, Long storeId) {
+    public CommonResponseDto updateStoreContent(String content, Long storeId, Long id) {
+        User user = userRepository.findById(id).orElseThrow(
+                () -> new CustomException(ExceptionCode.NOT_FOUND_USER)
+        );
+        if(!user.getRole().equals(SELLER)){
+            throw new CustomException(ExceptionCode.FORBIDDEN_YOUR_NOT_SELLER);
+        }
         Store store = findStore(storeId);
         store.setContent(content);
         return new CommonResponseDto(CommonResponseCode.STORE_UPDATE);
     }
 
     @Transactional
-    public CommonResponseDto updateStoreMinDeliveryPrice(Integer minDeliveryPrice, Long storeId) {
+    public CommonResponseDto updateStoreMinDeliveryPrice(Integer minDeliveryPrice, Long storeId, Long id) {
+        User user = userRepository.findById(id).orElseThrow(
+                () -> new CustomException(ExceptionCode.NOT_FOUND_USER)
+        );
+        if(!user.getRole().equals(SELLER)){
+            throw new CustomException(ExceptionCode.FORBIDDEN_YOUR_NOT_SELLER);
+        }
         Store store = findStore(storeId);
         store.setMinDeliveryPrice(minDeliveryPrice);
         return new CommonResponseDto(CommonResponseCode.STORE_UPDATE);
     }
 
     @Transactional
-    public CommonResponseDto updateStoreOpenStatus(Boolean openStatus, Long storeId) {
+    public CommonResponseDto updateStoreOpenStatus(Boolean openStatus, Long storeId, Long id) {
+        User user = userRepository.findById(id).orElseThrow(
+                () -> new CustomException(ExceptionCode.NOT_FOUND_USER)
+        );
+        if(!user.getRole().equals(SELLER)){
+            throw new CustomException(ExceptionCode.FORBIDDEN_YOUR_NOT_SELLER);
+        }
         Store store = findStore(storeId);
         store.setOpenStatus(openStatus);
         return new CommonResponseDto(CommonResponseCode.STORE_UPDATE);
     }
 
-    public List<StoreResponseDto> getStoreList() {
-        return storeRepository.findAll().stream().map(this::convertStoreResponseDto).toList();
+    // User 든 Seller든 모두 가능
+    public List<StoreResponseDto> getTotalStoreList(Long id) {
+        User user = userRepository.findById(id).orElseThrow(
+                () -> new CustomException(ExceptionCode.NOT_FOUND_USER)
+        );
+        return storeRepository.findAllByIsDeletedFalse().stream().map(this::convertStoreResponseDto).toList();
+    }
+    public List<StoreResponseDto> getStoreList(Long id) {
+        User user = userRepository.findById(id).orElseThrow(
+                () -> new CustomException(ExceptionCode.NOT_FOUND_USER)
+        );
+        if(!user.getRole().equals(SELLER)){
+            throw new CustomException(ExceptionCode.FORBIDDEN_YOUR_NOT_SELLER);
+        }
+        return storeRepository.findAllByUserIdAndIsDeletedFalse(id).stream().map(this::convertStoreResponseDto).toList();
     }
 
-
-    public StoreResponseDto getStore(Long storeId) {
+    public StoreResponseDto getStore(Long id, Long storeId) {
+        User user = userRepository.findById(id).orElseThrow(
+                () -> new CustomException(ExceptionCode.NOT_FOUND_USER)
+        );
         Store store = findStore(storeId);
+        if(store.isDeleted()){
+            throw new CustomException(ExceptionCode.NOT_FOUND_STORE);
+        }
         return convertStoreResponseDto(store);
     }
     private Store findStore(Long id) {
@@ -146,4 +213,5 @@ public class StoreService {
         storeResponseDto.setOpenStatus(store.isOpenStatus());
         return storeResponseDto;
     }
+
 }
